+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+### Added
+
+- Scrape logs from kube-system and giantswarm namespaces only.
+
 ## [0.0.3] - 2023-07-27
 
 ### Fixed